backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.0
python-multipart==0.0.9
librosa==0.10.2
numpy>=1.26.0
soundfile==0.12.1
scipy>=1.11.4
pydantic>=2.7.0,<3.0.0
python-dotenv>=1.0.0

